requirements.txt
Flask==2.0.2
Werkzeug==2.0.2
Jinja2==3.0.2
cassandra-driver==3.28.0
